refactor: extract scanner as a component
