+++ src/modules/bookScan/components/styles/bookScanner.style.js
