+++ src/redux/actions/index.js
@@ -1,3 +1,4 @@
 export * from './async/authenticationAsyncActions'
 export * from './sync/authenticationActions'
 export * from './sync/shoppingBagActions'
+export * from './async/bookActions'
